requirements.txt
streamlit
opencv-python
pytesseract
Pillow
rapidfuzz
pyyaml
numpy
